+++ CHANGES.rst
@@ -46,6 +46,13 @@ Changes
   ``fromUnicode``. See `issue 28
   <https://github.com/zopefoundation/zope.configuration/issues/28>`_.
 
+- Add ``ConfigurationMachine.pass_through_exceptions`` to allow
+  customizing the exceptions that
+  ``ConfigurationMachine.execute_actions`` wraps in a
+  ``ConfigurationExecutionError``. See `issue 10
+  <https://github.com/zopefoundation/zope.configuration/issues/10>`_.
+
+
 4.1.0 (2017-04-26)
 ------------------